requirements.txt
fastapi
uvicorn
pydantic
wandb
torch
ray[serve]
python-dotenv
requests



